Add support for importing in DSi NAND apps that are not DSiWare
They'll be considered internally "DSiWare" for purposes of save management and the movie header (although they aren't technically DSiWare, just system DSi NAND apps)

+++ src/BizHawk.Emulation.Cores/Consoles/Nintendo/NDS/LibMelonDS.cs
@@ -151,13 +151,13 @@ namespace BizHawk.Emulation.Cores.Consoles.Nintendo.NDS
 		public abstract bool SaveRamIsDirty();
 
 		[BizImport(CC)]
-		public abstract void ImportDSiWareSavs(IntPtr console, uint titleId);
+		public abstract void ImportDSiWareSavs(IntPtr console, ulong titleId);
 
 		[BizImport(CC)]
-		public abstract void ExportDSiWareSavs(IntPtr console, uint titleId);
+		public abstract void ExportDSiWareSavs(IntPtr console, ulong titleId);
 
 		[BizImport(CC)]
-		public abstract void DSiWareSavsLength(IntPtr console, uint titleId, out int publicSavSize, out int privateSavSize, out int bannerSavSize);
+		public abstract void DSiWareSavsLength(IntPtr console, ulong titleId, out int publicSavSize, out int privateSavSize, out int bannerSavSize);
 
 		[BizImport(CC)]
 		public abstract void GetRegs(IntPtr console, uint[] regs);
